+++ Controls-demo/PropertyGridNew/Editors/AfterEditorTemplate/AfterEditorTemplate.wml
@@ -1,14 +1,21 @@
-<div class="controlsDemo__wrapper controlsDemo__wrapper__afterEditorTemplate controls-Demo__wide-propertyGrid">
-    <Controls.propertyGrid:PropertyGrid
-            bind:editingObject="_editingObject"
-            typeDescription="{{_typeDescription}}"
-            captionPosition="left">
-        <ws:itemTemplate>
-            <ws:partial template="Controls/propertyGrid:ItemTemplate">
-                <ws:afterEditorTemplate>
-                    <Controls.buttons:Button buttonStyle="pale" iconSize="s" icon="icon-Question2" viewMode="functionalButton"/>
-                </ws:afterEditorTemplate>
-            </ws:partial>
-        </ws:itemTemplate>
-    </Controls.propertyGrid:PropertyGrid>
+<div class="controlsDemo__wrapper">
+    <div class="controlsDemo__flexRow">
+        <div class="controlsDemo__cell controlsDemo__mr4">
+            <Router.router:Reference attr:class="controls-text-label"
+                                     state="/Controls-demo/app/:app"
+                                     app="{{ 'Controls-demo/PropertyGridNew/Editors/AfterEditorTemplate/Render/Index' }}">
+                <a href="{{content.href}}">Render</a>
+            </Router.router:Reference>
+            <Controls-demo.PropertyGridNew.Editors.AfterEditorTemplate.Render.Index/>
+        </div>
+
+        <div class="controlsDemo__cell controlsDemo__mr4">
+            <Router.router:Reference attr:class="controls-text-label"
+                                     state="/Controls-demo/app/:app"
+                                     app="{{ 'Controls-demo/PropertyGridNew/Editors/AfterEditorTemplate/GridRender/Index' }}">
+                <a href="{{content.href}}">Grid Render</a>
+            </Router.router:Reference>
+            <Controls-demo.PropertyGridNew.Editors.AfterEditorTemplate.GridRender.Index/>
+        </div>
+    </div>
 </div>
